Refactor LDAP tests
Local tests run against OpenLDAP Docker container with sample data.

Public tests run against OpenLDAP server hosted by enterprising 3rd party ( https://www.forumsys.com/tutorials/integration-how-to/ldap/online-ldap-test-server).

AD tests run against AD server hosted in Azure.
